# document/flowpilot-rest-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: FlowPilot REST API
  version: 1.5.0
  license:
    name: Proprietary - Internal Use Only
    identifier: LicenseRef-Proprietary
  description: |
    FlowPilot 内网流程平台首版 REST API 契约。

    所有写操作由后端重新执行权限、状态与版本校验。可变资源使用整数
    `revision` 生成强 ETag；要求 `If-Match` 的接口在缺少请求头时返回 428，
    在版本过期时返回 412。创建及领域命令要求 `Idempotency-Key`，同一用户、
    路由作用域和键的相同请求会重放首次结果，不同请求体会返回 409。

    日期时间均为 ISO-8601 UTC。业务实体 ID 均由后端生成。审批、确认、字段
    修改、任务激活、流程状态、审计与邮件 Outbox 必须在同一业务事务中提交；
    SMTP 发送在事务提交后异步执行。Excel 文件只在浏览器中根据后端返回的
    已鉴权导出数据集生成，后端不生成、不保存 Excel 文件。

    浏览器只使用同源 HttpOnly 会话 Cookie，不保存访问令牌或业务数据。
servers:
  - url: /api/flowpilot/v1
    description: 与前端同源部署的内网 API
security:
  - cookieAuth: []
...
paths:
...
  /process-instances:
    get:
      tags: [ProcessInstances]
      operationId: listProcessInstances
      summary: 按当前用户数据范围分页查询流程实例
      description: dateFrom/dateTo 必填；前端默认今天至前 30 天且不可清空。“我的发起”传入当前用户、activeOnly=true 时返回本人全部未结束实例，不按日期截断。
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/Search"
        - name: definitionId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProcessInstanceStatus"
        - name: initiatorId
          in: query
          schema:
            type: string
            format: uuid
        - name: activeOnly
          in: query
          description: 为 true 时仅返回尚未完成或关闭的实例，供“我的发起”分页使用
          schema:
            type: boolean
        - name: currentNode
          in: query
          schema:
            type: string
        - name: dynamicFilters
          in: query
          style: deepObject
          explode: true
          description: 仅允许目标版本中 queryable=true 的稳定字段 ID
          schema:
            type: object
            additionalProperties: true
      responses:
        "200":
          description: 流程实例分页结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessInstancePage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [ProcessInstances]
      operationId: createProcessInstance
      summary: 提交表单并创建流程实例
      description: |
        单事务完成发布版本锁定、表单与附件引用保存、共享月度编号生成、首轮任务
        展开、实例计数、时间线和节点激活邮件 Outbox。页面预览不占用编号。
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessInstanceRequest"
      responses:
        "201":
          description: 实例已创建
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessInstanceDetailDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /me/workflow-tasks:
    get:
      tags: [WorkflowTasks]
      operationId: listMyWorkflowTasks
      summary: 查询任务中心中的我的待办或可代办任务
      description: 按流程实例分页，每项返回该实例下当前视图可见的任务数组与实例摘要；categories 返回忽略 definitionId 条件后的各流程完整数量汇总。tasks 中各任务通过 taskType 区分审批、自由协作和待重新提交，后两类没有审批节点。待办不按创建日期截断。
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: view
          in: query
          required: true
          schema:
            type: string
            enum: [pending, substitutable]
        - name: definitionId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 任务分页结果；每项包含服务端计算的 allowedActions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCenterItemPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /workflow-tasks/{taskId}:
    get:
      tags: [WorkflowTasks]
      operationId: getWorkflowTask
      summary: 获取任务及可处理动作
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: 任务详情
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTaskDetailDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /workflow-tasks/{taskId}/decision:
    post:
      tags: [WorkflowTasks]
      operationId: decideWorkflowTask
      summary: 提交通过、确认或驳回结果
      description: |
        仅接受 taskType=approval；自由协作和待重新提交任务调用本端点返回 409。
        以条件更新保证默认责任人与代办人员并发时首个成功提交生效。确认节点只接受
        confirm；审批节点接受 pass/reject。授权字段修改、结果、实际处理人、后续激活、
        并行取消、实例状态、时间线和 Outbox 原子保存。
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskDecisionRequest"
      responses:
        "200":
          description: 任务处理成功
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDecisionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /workflow-tasks/{taskId}/field-revisions:
    post:
      tags: [WorkflowTasks]
      operationId: reviseCompletedWorkflowTaskFields
      summary: 在不改变原审核结果的前提下继续修改授权字段
      description: |
        仅接受 taskType=approval；自由协作和待重新提交任务调用本端点返回 409。
        只允许原实际处理人或超级管理员；节点必须开启重复修改，原结果必须为通过或
        确认，实例不得处于驳回待处理或已关闭。每次请求追加一条不可变差异记录，不
        重复生成审核记录，也不反向重算已激活、跳过或完成的节点。
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviseTaskFieldsRequest"
      responses:
        "201":
          description: 字段已更新且差异已留痕
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviseTaskFieldsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowpilot_session
      description: 正式同源 HttpOnly 会话 Cookie，Path=/api/flowpilot；当前 HTTP 环境 Secure=false，启用 HTTPS 后必须为 true
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Search:
      name: q
      in: query
      description: 去除首尾空格的关键词
      schema:
        type: string
        maxLength: 100
    DateFrom:
      name: dateFrom
      in: query
      required: true
      description: 查询起始日期，前端默认今天之前 30 天
      schema:
        type: string
        format: date
    DateTo:
      name: dateTo
      in: query
      required: true
      description: 查询结束日期，前端默认今天；不得早于 dateFrom
      schema:
        type: string
        format: date
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: GET 返回的强 ETag，例如 `"12"`
      schema:
        type: string
        pattern: '^"[0-9]+"$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端生成的 UUID；同作用域至少保留 24 小时
      schema:
        type: string
        minLength: 16
        maxLength: 100
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    ETag:
      description: 由资源 revision 生成的强 ETag
      schema:
        type: string
        pattern: '^"[0-9]+"$'
...
  responses:
    BadRequest:
      description: 请求语法、查询范围或媒体格式参数错误
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: 未登录或会话失效
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: 缺少页面/动作权限、流程资格或数据范围
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: 资源不存在，或为避免越权信息泄露而按不存在处理
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: 当前领域状态不允许该动作、幂等键冲突或字段并发冲突
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    PreconditionFailed:
      description: If-Match 已过期
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    UnprocessableEntity:
      description: 领域字段校验未通过
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    PreconditionRequired:
      description: 缺少必需的 If-Match
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
...
    ProcessInstanceStatus:
      type: string
      enum: [reviewing, rejected-pending, completed, in-progress, closed]
    CreateProcessInstanceRequest:
      type: object
      additionalProperties: false
      required: [definitionId, formValues]
      properties:
        definitionId:
          type: string
          format: uuid
        formValues:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/JsonValue"
        copySourceInstanceId:
          type: string
          format: uuid
          description: 复制预填提交时传入来源已完成实例；后端只校验来源、权限和审计，不在提交前创建任何实例
        assigneeByNode:
          type: object
          additionalProperties:
            type: string
            format: uuid
        firstAssigneeId:
          type: string
          format: uuid
          description: 自由协作流程的首位受理人
        attachmentIds:
          type: array
          description: 与本次新增回复一起提交的暂存附件标识。
          uniqueItems: true
          items:
            type: string
            format: uuid
        attachmentIdsByField:
          type: object
          description: 按稳定表单字段标识关联已上传的临时附件；服务端在实例创建事务内转为正式引用。
          additionalProperties:
            type: array
            uniqueItems: true
            items:
              type: string
              format: uuid
...
    ProcessInstancePage:
      type: object
      required: [items, meta, categories]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProcessInstanceSummaryDto"
        meta:
          $ref: "#/components/schemas/PageMeta"
        categories:
          type: array
          description: 忽略 definitionId 条件、保留其余查询条件后的各流程完整数量汇总
          items:
            $ref: "#/components/schemas/TaskCenterFlowCategoryDto"
...
    ProcessInstanceDetailDto:
      allOf:
        - $ref: "#/components/schemas/ProcessInstanceSummaryDto"
        - type: object
          required: [formValues, fieldRevisions, attachments, participants, canClose, reviewProgress, tasks, timeline, freeTimeline]
          properties:
            formValues:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/JsonValue"
            fieldRevisions:
              type: object
              description: 每个稳定字段 ID 的当前并发版本
              additionalProperties:
                type: integer
                minimum: 0
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/AttachmentDto"
            participants:
              type: array
              description: 自由协作事项的发起人、当前受理人及历史参与人；审批流程为空数组。
              items:
                $ref: "#/components/schemas/UserRef"
            canTransferFree:
              type: boolean
              description: 当前登录人是否可以在此进行中的自由协作事项中变更受理人；非自由流程或非进行中事项为 false。
            canClose:
              type: boolean
              description: 后端根据当前会话的关闭动作权限和实例锁定版本的关闭流程权限组计算；超级管理员具有关闭资格。
            freeAssigneeCandidates:
              type: array
              description: 此自由协作流程锁定版本的受理权限组中当前有效且启用的用户；非自由流程为空数组。
              items:
                $ref: "#/components/schemas/UserRef"
            reviewProgress:
              type: array
              items:
                $ref: "#/components/schemas/ReviewProgressDto"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowTaskDto"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEventDto"
            freeTimeline:
              type: array
              items:
                $ref: "#/components/schemas/FreeTimelineEntryDto"
...
    TaskAction:
      type: string
      enum: [pass, confirm, reject]
...
    WorkflowTaskDetailDto:
      type: object
      additionalProperties: false
      required: [task, instance]
      properties:
        task:
          $ref: "#/components/schemas/TaskDto"
        instance:
          $ref: "#/components/schemas/ProcessInstanceSummaryDto"
    TaskCenterItemPage:
      type: object
      required: [items, meta, categories]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskCenterItemDto"
        meta:
          $ref: "#/components/schemas/PageMeta"
        categories:
          type: array
          description: 忽略 definitionId 条件、保留当前任务视图和搜索条件后的各流程完整数量汇总
          items:
            $ref: "#/components/schemas/TaskCenterFlowCategoryDto"
...
    TaskDecisionRequest:
      type: object
      additionalProperties: false
      required: [action, fieldValues, baseFieldRevisions]
      properties:
        action:
          $ref: "#/components/schemas/TaskAction"
        comment:
          type: string
          maxLength: 2000
          description: 驳回必填；通过意见和确认说明选填
        fieldValues:
          type: object
          description: 只允许节点 editableFieldIds 中的字段；不接受整份表单覆盖
          additionalProperties:
            $ref: "#/components/schemas/JsonValue"
        baseFieldRevisions:
          $ref: "#/components/schemas/BaseFieldRevisions"
        attachmentIdsByField:
          type: object
          description: 本次审核字段修改关联的附件标识，与任务结果在同一事务内提交。
          additionalProperties:
            type: array
            uniqueItems: true
            items:
              type: string
              format: uuid
    TaskDecisionResponse:
      type: object
      required: [instance, task, activatedTaskIds, cancelledTaskIds]
      properties:
        instance:
          $ref: "#/components/schemas/ProcessInstanceDetailDto"
        task:
          $ref: "#/components/schemas/WorkflowTaskDto"
        activatedTaskIds:
          type: array
          items:
            type: string
            format: uuid
        cancelledTaskIds:
          type: array
          items:
            type: string
            format: uuid
...